# ai-service/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "江苏 13 市 AI 稽核助手",
    "description": "江苏 13 市 AI 稽核助手 REST API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/memories": {
      "get": {
        "tags": [
          "memories"
        ],
        "summary": "List Memories",
        "description": "查询当前城市全部长期记忆，供独立记忆管理页面使用。",
        "operationId": "list_memories_api_v1_memories_get",
        "parameters": [
          {
            "name": "memory_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "audit_case",
                    "correction_memory"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Memory Type"
            }
          },
          {
            "name": "memory_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "active",
                    "paused",
                    "invalidated"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Memory Status"
            }
          },
          {
            "name": "site_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memories/{memory_type}/{memory_id}/status": {
      "patch": {
        "tags": [
          "memories"
        ],
        "summary": "Update Memory Status",
        "description": "暂停或恢复一条未确认错误的记忆。",
        "operationId": "update_memory_status_api_v1_memories__memory_type___memory_id__status_patch",
        "parameters": [
          {
            "name": "memory_type",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "audit_case",
                "correction_memory"
              ],
              "type": "string",
              "title": "Memory Type"
            }
          },
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          },
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memories/{memory_type}/{memory_id}/flags": {
      "post": {
        "tags": [
          "memories"
        ],
        "summary": "Flag Memory",
        "description": "独立标错任意历史案例或人工纠错，并立即暂停其RAG使用。",
        "operationId": "flag_memory_api_v1_memories__memory_type___memory_id__flags_post",
        "parameters": [
          {
            "name": "memory_type",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "audit_case",
                "correction_memory"
              ],
              "type": "string",
              "title": "Memory Type"
            }
          },
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          },
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryFlagCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryFlagView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory-flags/{flag_id}": {
      "patch": {
        "tags": [
          "memories"
        ],
        "summary": "Resolve Memory Flag",
        "description": "将标错判为误标并恢复，或确认错误并使旧版本失效。",
        "operationId": "resolve_memory_flag_api_v1_memory_flags__flag_id__patch",
        "parameters": [
          {
            "name": "flag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Flag Id"
            }
          },
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryFlagResolve"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryFlagView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memories/{memory_type}/{memory_id}/revisions": {
      "post": {
        "tags": [
          "memories"
        ],
        "summary": "Revise Memory",
        "description": "修改已确认错误的记忆，保留旧快照并生成重新启用的新版本。",
        "operationId": "revise_memory_api_v1_memories__memory_type___memory_id__revisions_post",
        "parameters": [
          {
            "name": "memory_type",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "audit_case",
                "correction_memory"
              ],
              "type": "string",
              "title": "Memory Type"
            }
          },
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          },
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryRevisionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryRevisionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analysis-runs/{run_id}/reports": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Create Report",
        "description": "把完成的AI草稿保存为可编辑报告；重复调用返回同一资源。",
        "operationId": "create_report_api_v1_analysis_runs__run_id__reports_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{report_id}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Report",
        "description": "读取报告当前版本、状态和版本历史。",
        "operationId": "get_report_api_v1_reports__report_id__get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Report Id"
            }
          },
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "reports"
        ],
        "summary": "Update Report",
        "description": "保存一个新的完整报告版本。",
        "operationId": "update_report_api_v1_reports__report_id__patch",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Report Id"
            }
          },
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MemoryFlagCreate": {
        "properties": {
          "flag_type": {
            "type": "string",
            "enum": [
              "wrong_reason",
              "wrong_site",
              "wrong_evidence",
              "wrong_scope",
              "duplicate",
              "outdated",
              "other"
            ],
            "title": "Flag Type"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 2,
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "flag_type",
          "description"
        ],
        "title": "MemoryFlagCreate",
        "description": "业务员标记既有记忆错误时提交的原因。"
      },
      "MemoryFlagResolve": {
        "properties": {
          "resolution": {
            "type": "string",
            "enum": [
              "dismissed",
              "invalidated"
            ],
            "title": "Resolution"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "resolution"
        ],
        "title": "MemoryFlagResolve",
        "description": "复核标错：误标则恢复，确认错误则永久隔离旧版本。"
      },
      "MemoryFlagView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "memory_type": {
            "type": "string",
            "enum": [
              "audit_case",
              "correction_memory"
            ],
            "title": "Memory Type"
          },
          "memory_id": {
            "type": "string",
            "title": "Memory Id"
          },
          "city_code": {
            "type": "string",
            "title": "City Code"
          },
          "flag_type": {
            "type": "string",
            "title": "Flag Type"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "previous_memory_status": {
            "type": "string",
            "enum": [
              "active",
              "paused",
              "invalidated"
            ],
            "title": "Previous Memory Status"
          },
          "resolution_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Action"
          },
          "resolution_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Note"
          },
          "reported_at": {
            "type": "string",
            "format": "date-time",
            "title": "Reported At"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "memory_type",
          "memory_id",
          "city_code",
          "flag_type",
          "description",
          "status",
          "previous_memory_status",
          "resolution_action",
          "resolution_note",
          "reported_at",
          "resolved_at"
        ],
        "title": "MemoryFlagView",
        "description": "标错记录响应。"
      },
      "MemoryList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MemoryView"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "MemoryList",
        "description": "当前城市的统一记忆列表。"
      },
      "MemoryRevisionCreate": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 2,
            "title": "Reason"
          },
          "reason_category": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Reason Category"
          },
          "conditions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Conditions"
          },
          "evidence_element_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array",
                "maxItems": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Element Ids"
          },
          "change_reason": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 2,
            "title": "Change Reason"
          }
        },
        "type": "object",
        "required": [
          "reason",
          "reason_category",
          "change_reason"
        ],
        "title": "MemoryRevisionCreate",
        "description": "修改错误记忆并重新启用时允许替换的业务内容。"
      },
      "MemoryRevisionView": {
        "properties": {
          "memory": {
            "$ref": "#/components/schemas/MemoryView"
          },
          "replaced_version": {
            "type": "integer",
            "title": "Replaced Version"
          }
        },
        "type": "object",
        "required": [
          "memory",
          "replaced_version"
        ],
        "title": "MemoryRevisionView",
        "description": "修订后的新记忆以及旧版本号。"
      },
      "MemoryStatusUpdate": {
        "properties": {
          "memory_status": {
            "type": "string",
            "enum": [
              "active",
              "paused"
            ],
            "title": "Memory Status"
          }
        },
        "type": "object",
        "required": [
          "memory_status"
        ],
        "title": "MemoryStatusUpdate",
        "description": "日常暂停或恢复操作；无效记忆必须修订，不能直接恢复。"
      },
      "MemoryView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "memory_type": {
            "type": "string",
            "enum": [
              "audit_case",
              "correction_memory"
            ],
            "title": "Memory Type"
          },
          "city_code": {
            "type": "string",
            "title": "City Code"
          },
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "site_name": {
            "type": "string",
            "title": "Site Name"
          },
          "source_label": {
            "type": "string",
            "title": "Source Label"
          },
          "process_status": {
            "type": "string",
            "title": "Process Status"
          },
          "memory_status": {
            "type": "string",
            "enum": [
              "active",
              "paused",
              "invalidated"
            ],
            "title": "Memory Status"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "reason_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Category"
          },
          "conditions": {
            "items": {},
            "type": "array",
            "title": "Conditions"
          },
          "evidence_element_ids": {
            "items": {},
            "type": "array",
            "title": "Evidence Element Ids"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "open_flag_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open Flag Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "memory_type",
          "city_code",
          "site_id",
          "site_name",
          "source_label",
          "process_status",
          "memory_status",
          "reason",
          "reason_category",
          "conditions",
          "evidence_element_ids",
          "version",
          "source_id",
          "created_at",
          "updated_at"
        ],
        "title": "MemoryView",
        "description": "把两张来源表转换成页面可统一展示的记忆资源。"
      },
...
      "ReportUpdate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Title"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/ReportSectionEdit"
            },
            "type": "array",
            "maxItems": 6,
            "minItems": 6,
            "title": "Sections"
          },
          "change_summary": {
            "type": "string",
            "maxLength": 500,
            "minLength": 2,
            "title": "Change Summary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "sections",
          "change_summary"
        ],
        "title": "ReportUpdate",
        "description": "保存一个完整的新版本，避免局部更新留下章节缺失。"
      },
...
      "ReportView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "city_code": {
            "type": "string",
            "title": "City Code"
          },
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "analysis_run_id": {
            "type": "string",
            "title": "Analysis Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "in_review",
              "returned",
              "approved"
            ],
            "title": "Status"
          },
          "current_version": {
            "type": "integer",
            "title": "Current Version"
          },
          "approved_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved Version"
          },
          "review_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Note"
          },
          "current": {
            "$ref": "#/components/schemas/ReportVersionView"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/ReportVersionSummary"
            },
            "type": "array",
            "title": "Versions"
          },
          "content_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Url"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "city_code",
          "site_id",
          "task_id",
          "analysis_run_id",
          "status",
          "current_version",
          "approved_version",
          "review_note",
          "current",
          "versions",
          "content_url",
          "created_by",
          "reviewed_by",
          "created_at",
          "updated_at",
          "reviewed_at"
        ],
        "title": "ReportView"
      },
...
    }
  }
}